+++ repos/paella-extra-plugins/changelog.md
@@ -3,3 +3,7 @@
 - 2.5.0:
     * new paella-extra-plugins
     * paella-onboarding-plugin and paella-cookieconsient-plugin moved to paella-extras-plugin
+    * new plugin: es.upv.paella.fileContent
+    * new plugin: es.upv.paella.version
+    * new plugin: es.upv.paella.relatedVideos
+    * new plugin: es.upv.paella.keyboardShortcuts